# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Shalinle application.",
        "title": "Shalinle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/stops": {
            "get": {
                "description": "Get a list of all stops",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stops"
                ],
                "summary": "Get all stops",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/stops.StopDetailResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/stops.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stops/{id}": {
            "get": {
                "description": "Get a stop by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stops"
                ],
                "summary": "Get a stop by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stops.StopDetailResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/stops.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/stops.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "stops.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "stops.StopDetailResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}